requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anytree==2.13.0
arrow==1.3.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autocorrect==2.6.1
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.24.0
azure-mgmt-core==1.6.0
azure-storage-blob==12.19.0
azureml-mlflow==1.60.0.post1
backoff==2.2.1
beautifulsoup4==4.13.5
bert-score==0.3.13
bioc==2.1
blessed==1.21.0
boto3==1.40.25
botocore==1.40.25
cachetools==5.5.2
causal-conv1d @ file:///cusp-data-efa/peihaow/jz/IBMamba/mamba_ssm/causal-conv1d
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
conllu==4.5.3
contourpy==1.3.2
croniter==1.4.1
cryptography==45.0.7
cycler==0.12.1
databricks-sdk==0.65.0
DataProperty==1.1.0
datasets==3.6.0
dateutils==0.6.12
debugpy==1.8.17
deepdiff==6.7.1
deepspeed==0.17.5
Deprecated==1.2.18
dill==0.3.8
docopt==0.6.2
docstring_parser==0.17.0
dropout_layer_norm @ file:///cusp-data-efa/peihaow/jz/IBMamba/flash-attention/csrc/layer_norm
editdistance==0.8.1
editor==1.6.6
einops==0.8.1
evaluate==0.4.5
exceptiongroup==1.3.0
filelock==3.13.1
fire==0.7.1
flair==0.15.1
flash-linear-attention @ file:///cusp-data-efa/peihaow/jz/IBMamba/flash-linear-attention
flash_attn==2.8.3
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2023.12.2
ftfy==6.3.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.9
hjson==3.1.0
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
inquirer==3.4.1
intervaltree==3.1.0
isodate==0.7.2
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.2
jsonargparse==4.41.0
jsonlines==4.0.0
jsonpickle==4.1.1
keras==3.11.3
kiwisolver==1.4.9
langdetect==1.0.9
language_tool_python==2.9.4
lemminflect==0.2.3
libclang==18.1.1
lightning==2.1.2
lightning-cloud==0.5.52
lightning-utilities==0.15.2
lm_eval==0.4.2
lru-dict==1.3.0
lxml==6.0.1
mamba-ssm @ file:///cusp-data-efa/peihaow/jz/IBMamba/mamba_ssm
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mbstrdecoder==1.1.4
mdurl==0.1.2
ml_dtypes==0.5.3
mlflow-skinny==2.22.2
more-itertools==10.8.0
mpld3==0.5.11
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.1.1
msrest==0.7.1
multidict==6.6.4
multiprocess==0.70.16
namex==0.1.0
networkx==3.3
ninja==1.13.0
nltk==3.9.1
num2words==0.5.14
numexpr==2.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
OpenHowNet==2.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
optree==0.17.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.2
pathvalidate==3.3.1
peft==0.17.1
pillow==11.3.0
pinyin==0.4.0
platformdirs==4.4.0
pluggy==1.6.0
portalocker==3.2.0
pptree==3.1
promptbench==0.0.4
propcache==0.3.2
protobuf==6.32.1
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.4
PySocks==1.7.1
pytablewriter==1.2.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytorch_revgrad==0.2.0
pytz==2025.2
PyYAML==6.0.2
readchar==4.2.1
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rotary_emb @ file:///cusp-data-efa/peihaow/jz/IBMamba/flash-attention/csrc/rotary
rouge_score==0.1.2
rsa==4.9.1
runs==1.2.2
s3transfer==0.13.1
sacrebleu==2.5.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
segtok==1.5.11
sentencepiece==0.2.1
sentry-sdk==2.37.0
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.8
sqlitedict==2.1.0
sqlparse==0.5.3
starsessions==1.3.0
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
termcolor==3.1.0
terminaltables==3.1.10
textattack==0.3.10
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
torch==2.5.1+cu121
torchdata==0.11.0
torchmetrics==1.8.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.2
transformers==4.44.2
triton-nightly==3.0.0.post20240716052845
trl==0.11.4
typeguard==4.4.4
typepy==1.3.4
types-python-dateutil==2.9.0.20250822
typeshed_client==2.8.2
typing_extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
Wikipedia-API==0.8.1
word2number==1.1
wrapt==1.17.3
xentropy_cuda_lib @ file:///cusp-data-efa/peihaow/jz/IBMamba/flash-attention/csrc/xentropy
xformers==0.0.29.post1
xmod==1.8.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
